requirements.txt
langchain-openai
langchain
python-dotenv
langchain_community
ipykernel
langchain_core
streamlit
bs4
arxiv
pymupdf
wikipedia
langchain-text-splitters
langchain-openai
chromadb
sentence_transformers
langchain_huggingface
faiss-cpu
langchain_chroma
duckdb
pandas
openai
langchain_groq
pypdf
duckduckgo-search
sqlalchemy
mysql-connector-python
validators==0.28.1
youtube_transcript_api
unstructured
pytube
youtube-transcript-api
numexpr
huggingface_hub
cassio
datasets
tiktoken
PyPDF2
bitsandbytes
transformers
accelerate
gradio